backend/django/requirements.txt
asgiref==3.5.0
Django==4.0.3
djangorestframework==3.13.1
joblib==1.1.0
mysqlclient==2.1.0
numpy==1.22.2
pandas==1.4.1
python-dateutil==2.8.2
pytz==2021.3
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
sqlparse==0.4.2
threadpoolctl==3.1.0
tzdata==2021.5
